refactor: generateLottoNumbers 함수 내부 메서드로 수정
